% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.neighboring.probes.R
\name{get.neighboring.probes}
\alias{get.neighboring.probes}
\title{Get Neighboring Probes}
\usage{
get.neighboring.probes(
  X,
  Y,
  chr,
  max.dist,
  control.arms = TRUE,
  remove.duplicates = TRUE
)
}
\value{
A list.
}
\description{
Get neighboring probes.
}
