% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{order.feature.info}
\alias{order.feature.info}
\title{Order the gene information table by chromosomal locations.}
\usage{
order.feature.info(feature.info)
}
\arguments{
\item{feature.info}{A data frame containing at least the following fields:
geneName, chr, and loc.}
}
\value{
An ordered data frame.
}
\description{
Order the gene information table by chromosomal locations. Removes genes
with no location information.
}
\examples{

## NOT RUN
#feature.info.ordered <- order.feature.info(feature.info)

}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("pint")
}
\keyword{utilities}

